requirements.txt
google-cloud-storage==2.15.0
google-cloud-bigquery==3.18.0
pyspark==3.5.1
pandas==2.2.0
google-api-core==2.17.1
prefect==2.16.2
prefect-gcp==0.5.6
kaggle
